+++ lib/src/dots_indicator.dart
@@ -7,12 +7,16 @@ class DotsIndicator extends StatelessWidget {
   final int dotsCount;
   final int position;
   final DotsDecorator decorator;
+  final Axis axis;
+  final bool reversed;
 
-  DotsIndicator({
+  const DotsIndicator({
     Key key,
     @required this.dotsCount,
     this.position = 0,
     this.decorator = const DotsDecorator(),
+    this.axis = Axis.horizontal,
+    this.reversed = false,
   })  : assert(dotsCount != null && dotsCount > 0),
         assert(position != null && position >= 0),
         assert(decorator != null),
@@ -22,24 +26,34 @@ class DotsIndicator extends StatelessWidget {
         ),
         super(key: key);
 
-  Widget _buildDot(int i) {
-    final color = (i == position) ? decorator.activeColor : decorator.color;
-    final size = (i == position) ? decorator.activeSize : decorator.size;
-    final shape = (i == position) ? decorator.activeShape : decorator.shape;
+  Widget _buildDot(int index) {
+    final isCurrent = index == position;
+    final size = isCurrent ? decorator.activeSize : decorator.size;
 
     return Container(
       width: size.width,
       height: size.height,
       margin: decorator.spacing,
-      decoration: ShapeDecoration(color: color, shape: shape),
+      decoration: ShapeDecoration(
+        color: isCurrent ? decorator.activeColor : decorator.color,
+        shape: isCurrent ? decorator.activeShape : decorator.shape,
+      ),
     );
   }
 
   @override
   Widget build(BuildContext context) {
+    final dotsList = List<Widget>.generate(dotsCount, _buildDot);
+
+    if (axis == Axis.vertical) {
+      return Column(
+        mainAxisSize: MainAxisSize.min,
+        children: reversed == true ? dotsList.reversed.toList() : dotsList,
+      );
+    }
     return Row(
       mainAxisSize: MainAxisSize.min,
-      children: List<Widget>.generate(dotsCount, _buildDot),
+      children: reversed == true ? dotsList.reversed.toList() : dotsList,
     );
   }
 }